Separate `object_hook` into own module.
