+++ blaze/utils.py
@@ -8,22 +8,23 @@ import os
 import re
 
 try:
-    from cytoolz import nth, unique, concat, first, drop, curry
+    from cytoolz import nth, unique, concat
 except ImportError:
-    from toolz import nth, unique, concat, first, drop, curry
+    from toolz import nth, unique, concat
 
-from datashape import dshape, Mono, DataShape
+from datashape import Mono, DataShape
 import numpy as np
 # these are used throughout blaze, don't remove them
 from odo.utils import tmpfile, filetext, filetexts, raises, keywords, ignoring
 import pandas as pd
 import psutil
 import sqlalchemy as sa
-import werkzeug.exceptions as wz_ex
 
 # Imports that replace older utils.
-from .compatibility import map, zip, PY2, builtins, reduce
+from .compatibility import map, zip
 from .dispatch import dispatch
+
+from .object_hook import object_hook
 
 
 def nth_list(n, seq):
@@ -201,129 +202,6 @@ def json_dumps(f):
     # object_hook, catch anything that is not pandas_numpy
     fcn = ".".join([f.__module__, f.__name__])
     return {'__!numpy_pandas_function': fcn}
-
-
-# dict of converters. This is stored as a default arg to object hook for
-# performance because this function is really really slow when unpacking data.
-# This is a mutable default but it is not a bug!
-_converters = {}
-
-
-if PY2:
-    _keys = dict.keys
-else:
-    def _keys(d, _dkeys=dict.keys, _list=list):
-        return _list(_dkeys(d))
-
-
-def object_hook(ob,
-                # Cached for performance. Forget these exist.
-                _len=len,
-                _keys=_keys,
-                _first_three_chars=np.s_[:3],
-                _converters=_converters):
-    """Convert a json object dict back into a python object.
-
-    This looks for our objects that have encoded richer representations
-    with a ``__!{type}`` key.
-
-    Parameters
-    ----------
-    ob : dict
-        The raw json parsed dictionary.
-
-    Returns
-    -------
-    parsed : any
-        The richer form of the object.
-
-    Notes
-    -----
-    The types that this reads can be extended with the ``register`` method.
-    For example:
-
-    >>> class MyList(list):
-    ...     pass
-    >>> @object_hook.register('MyList')
-    ... def _parse_my_list(ob):
-    ...     return MyList(ob)
-
-    Register can also be called as a function like:
-    >>> a = object_hook.register('frozenset', frozenset)
-    >>> a is frozenset
-    True
-    """
-    if _len(ob) != 1:
-        return ob
-
-    key = _keys(ob)[0]
-    if key[_first_three_chars] != '__!':
-        return ob
-
-    return _converters[key](ob[key])
-
-
-@curry
-def register(typename, converter, converters=_converters):
-    converters['__!' + typename] = converter
-    return converter
-object_hook.register = register
-del register
-
-object_hook._converters = _converters  # make this accesible for debugging
-del _converters
-del _keys  # captured by default args
-
-
-object_hook.register('datetime', pd.Timestamp)
-object_hook.register('frozenset', frozenset)
-object_hook.register('datashape', dshape)
-
-
-def builtins_function_from_str(f):
-    if f in ("eval", "exec"):
-        raise wz_ex.Forbidden("cannot invoke eval or exec")
-
-    return getattr(builtins, f)
-
-object_hook.register('builtin_function', builtins_function_from_str)
-
-
-def numpy_pandas_function_from_str(f):
-    """
-    reconstruct function from string representation
-    """
-    if f.startswith("numpy"):
-        mod = np
-
-    elif f.startswith("pandas"):
-        mod = pd
-
-    else:
-        raise wz_ex.NotImplemented("accepts numpy/pandas/builtin funcs only")
-
-    fcn = reduce(getattr, f.split('.')[1:], mod)
-    return fcn
-
-
-object_hook.register('numpy_pandas_function', numpy_pandas_function_from_str)
-
-
-@object_hook.register('mono')
-def _read_mono(m):
-    return dshape(m).measure
-
-
-@object_hook.register('timedelta')
-def _read_timedelta(ds):
-    return datetime.timedelta(seconds=ds)
-
-
-@object_hook.register('bytes')
-def _read_bytes(bs):
-    if not isinstance(bs, bytes):
-        bs = bs.encode('latin1')
-    return bs
 
 
 def normalize(s):
